+++ src/asserts/images/decline.png
