+++ src/asserts/wrappers/Error.js
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+const Wrapper = styled.main`
+    background: #94bdff;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    h1 {
+        font-size: 10rem;
+    }
+    h3 {
+        text-transform: none;
+        margin-bottom: 2rem;
+    }
+`
+
+export default Wrapper;
