+++ src/pages/Error.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
+import Wrapper from '../asserts/wrappers/Error'
 
 const Error = () => {
     return (
@@ -16,19 +17,6 @@ const Error = () => {
     );
 }
 
-const Wrapper = styled.main`
-    background: #94bdff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    h1 {
-        font-size: 10rem;
-    }
-    h3 {
-        text-transform: none;
-        margin-bottom: 2rem;
-    }
-`
+
 
 export default Error;
